fix: refactor cluster classes
