+++ src/commands/cluster/handlers.ts
@@ -27,14 +27,21 @@ import {patchInject} from "../../core/container_helper.js";
 import {K8} from "../../core/k8.js";
 import {CommandHandler} from "../../core/command_handler.js";
 import {LocalConfig} from "../../core/config/local_config.js";
+import type {CommandFlag} from "../../types/flag_types.js";
+import {ConfigManager} from "../../core/config_manager.js";
+import {ChartManager} from "../../core/chart_manager.js";
 
 @injectable()
 export class ClusterCommandHandlers extends CommandHandler {
+  protected readonly _configMaps = new Map<string, any>();
+
   constructor(
       @inject(ClusterCommandTasks) private readonly tasks: ClusterCommandTasks,
       @inject(RemoteConfigManager) private readonly remoteConfigManager: RemoteConfigManager,
       @inject(LocalConfig) private readonly localConfig: LocalConfig,
       @inject(K8) private readonly k8: K8,
+      @inject(ConfigManager) private readonly configManager: ConfigManager,
+      @inject(ChartManager) private readonly chartManager: ChartManager,
   ) {
       super();
 
@@ -42,6 +49,8 @@ export class ClusterCommandHandlers extends CommandHandler {
       this.remoteConfigManager = patchInject(remoteConfigManager, RemoteConfigManager, this.constructor.name);
       this.k8 = patchInject(k8, K8, this.constructor.name);
       this.localConfig = patchInject(localConfig, LocalConfig, this.constructor.name);
+      this.configManager = patchInject(configManager, ConfigManager, this.constructor.name);
+      this.chartManager = patchInject(chartManager, ChartManager, this.constructor.name);
   }
 
   async connect(argv: any) {
@@ -53,7 +62,7 @@ export class ClusterCommandHandlers extends CommandHandler {
         this.setupHomeDirectoryTask(),
         this.localConfig.promptLocalConfigTask(this.k8),
         this.tasks.selectContext(),
-        ListrRemoteConfig.loadRemoteConfig(this.parent, argv),
+        ListrRemoteConfig.loadRemoteConfig(this.remoteConfigManager, argv),
         this.tasks.readClustersFromRemoteConfig(argv),
         this.tasks.updateLocalConfig(),
       ],
@@ -153,4 +162,18 @@ export class ClusterCommandHandlers extends CommandHandler {
     }
     return true;
   }
+
+    close(): Promise<void> {
+        // no-op
+        return Promise.resolve();
+    }
+
+    // Config related methods:
+    getConfig(configName: string, flags: CommandFlag[], extraProperties: string[] = []): object {
+        return helpers.getConfig(this.configManager, this._configMaps, configName, flags, extraProperties);
+    }
+
+    getUnusedConfigs(configName: string): string[] {
+        return this._configMaps.get(configName).getUnusedConfigs();
+    }
 }
